requirements.txt
librosa==0.10.1
numpy==1.25.2
sounddevice==0.4.6
matplotlib==3.8.0
requests==2.31.0
torch==2.1.1  # For model conversion if needed
torchaudio==2.1.1
silero-vad==5.1.2
openai-whisper==20231117
